pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "langgraph>=0.5.1",
    "langsmith>=0.4.4",
    "notebook>=7.4.4",
    "python-dotenv>=1.1.1",
]
